% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter_data.R
\docType{data}
\name{epiparameter_data}
\alias{epiparameter_data}
\title{epiparameter database}
\format{
\subsection{\code{epiparameter_data}}{

An epidist object, which is a list:
\describe{
See details in \code{?epiparameter::epidist_db()}
}
}
}
\source{
\url{https://github.com/epiverse-trace/epiparameter}
}
\usage{
epiparameter_data
}
\description{
A re-exported database of epidemiological parameters from the
epiparameter package for internal use.
}
\keyword{datasets}
