% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{epiparameterDashboard-package}
\alias{epiparameterDashboard}
\alias{epiparameterDashboard-package}
\title{epiparameterDashboard: Explore The Database Of Epidemiological Parameters Provided By The epiparameterDashboard And Epireview R Packages}
\description{
Explore The Database Of Epidemiological Parameters Provided By The epiparameterDashboard And Epireview R Packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/epiparameterDashboard/}
  \item \url{https://epiverse-trace.github.io/epiparameterDashboard/}
  \item Report bugs at \url{https://github.com/epiverse-trace/epiparameterDashboard/issues}
}

}
\author{
\strong{Maintainer}: James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID})

Authors:
\itemize{
  \item Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID})
}

}
\keyword{internal}
