% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{epiparameterDashboard}
\alias{epiparameterDashboard}
\title{Epiparameter Dashboard}
\usage{
epiparameterDashboard(...)
}
\arguments{
\item{...}{Not used}
}
\value{
A shiny app that visualises epidemiological parameters.
}
\description{
Epiparameter Dashboard
}
\examples{
\dontrun{
epiparameterDashboard()
}
}
