% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{epiparameterDashboard-package}
\alias{epiparameterDashboard}
\alias{epiparameterDashboard-package}
\title{epiparameterDashboard: A Dashboard for Visualising Epidemiological Parameters}
\description{
A Dashboard for Visualising Epidemiological Parameters.
}
\author{
\strong{Maintainer}: James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID})

}
\keyword{internal}
